+++ packages/workspace-fusion/src/lib/utils/dataSource/index.ts
@@ -1,0 +1,1 @@
+export { addDataSource } from './addDataSource';
