+++ packages/workspace-fusion/src/lib/utils/fusionTabOrder/index.ts
@@ -1,0 +1,1 @@
+export { sortFusionTabs } from './fusionTabOrder';
