+++ packages/workspace-fusion/src/lib/utils/garden/index.ts
@@ -1,0 +1,1 @@
+export { addGarden } from './addGarden';
