+++ packages/workspace-fusion/src/lib/utils/index.ts
@@ -1,7 +1,9 @@
-export * from './addDataSource';
-export * from './addGrid';
-export * from './addGarden';
-export * from './addSidesheet';
-export * from './addStatusBar';
+export * from './dataSource';
+export * from './fusionTabOrder';
+export * from './garden';
+export * from './grid';
+export * from './sidesheet';
+export * from './statusBar';
 export * from './addCustomTab';
 export * from './createFusionWorkspace';
+export * from './viewController';
