+++ packages/workspace-fusion/src/lib/utils/sidesheet/index.ts
@@ -1,0 +1,1 @@
+export { addSidesheet } from './addSidesheet';
