+++ packages/workspace-fusion/src/lib/utils/statusBar/index.ts
@@ -1,0 +1,1 @@
+export { addStatusBar } from './addStatusBar';
